requirements.txt
dnspython
Flask
gunicorn
pymongo[srv]
python-dotenv